Move get_next_event into method of EventDetector
